% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_grid_sc.R
\name{facet_grid_sc}
\alias{facet_grid_sc}
\title{Lay out panels in a grid with different scales}
\usage{
facet_grid_sc(rows = NULL, cols = NULL, scales = "fixed",
  space = "fixed", shrink = TRUE, labeller = "label_value",
  as.table = TRUE, switch = NULL, drop = TRUE, margins = FALSE,
  facets = NULL)
}
\arguments{
\item{rows}{A set of variables or expressions quoted by
\code{\link[=vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}

\item{cols}{A set of variables or expressions quoted by
\code{\link[=vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}

\item{scales}{A list of two elements (`x` and `y`). Each element can be either
`"fixed"` (scale limits shared across facets), `"free"` (with varying limits per facet), or
 a named list, with a different scale for each facet value. Previous scale values
 (`"fixed"`, `"free_x"`, `"free_y"`, `"free"` are accepted but soft-deprecated).}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the
y scale; if \code{"free_x"} their width will be proportional to the
length of the x scale; or if \code{"free"} both height and width will
vary.  This setting has no effect unless the appropriate scales also vary.}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with formulae of the type \code{~cyl + am}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[=labeller]{labeller()}}. See
\code{\link[=label_value]{label_value()}} for more details and pointers to other
options.}

\item{as.table}{If \code{TRUE}, the default, the facets are laid out like
a table with highest values at the bottom-right. If \code{FALSE}, the
facets are laid out like a plot with the highest value at the top-right.}

\item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}

\item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}

\item{margins}{Either a logical value or a character
vector. Margins are additional facets which contain all the data
for each of the possible values of the faceting variables. If
\code{FALSE}, no additional facets are included (the
default). If \code{TRUE}, margins are included for all faceting
variables. If specified as a character vector, it is the names of
variables for which margins are to be created.}

\item{facets}{This argument is soft-deprecated, please use \code{rows}
and \code{cols} instead.}
}
\description{
`facet_grid_sc` is a variant of `facet_grid`
}
\examples{
library(ggplot2)
library(scales)
# Custom scales per facet:
 mydf <- data.frame(
   Subject = rep(c("A", "B", "C", "D"), each = 3),
   Magnitude = rep(c("SomeValue", "Percent", "Scientific"), times = 4),
   Value=c(c(170,0.6,2.7E-4),
           c(180, 0.8, 2.5E-4),
           c(160, 0.71, 3.2E-4),
           c(159, 0.62, 3E-4)))

 scales_y <- list(
   Percent = scale_y_continuous(labels=percent_format()),
   SomeValue = scale_y_continuous(),
   Scientific = scale_y_continuous(labels=scientific_format())
 )

 ggplot(mydf) +
   geom_point(aes(x=Subject, y=Value)) +
   facet_grid_sc(rows = vars(Magnitude), scales = list(y = scales_y))

}
