% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_grid_sc.R
\docType{data}
\name{FacetGridScales}
\alias{FacetGridScales}
\title{ggproto facet}
\format{An object of class \code{FacetGridScales} (inherits from \code{FacetGrid}, \code{Facet}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
FacetGridScales
}
\description{
ggproto facet
}
\keyword{datasets}
